TST: - split up tests for Hermite functions into dedicated test files to be more flexible
