+++ tests/test_hermite_functions/__init__.py
@@ -1,0 +1,4 @@
+"""
+This test suite implements the tests for the module :mod:`hermite_functions`.
+
+"""
